+++ subgrade-sdk/src/main/java/com/dwarfeng/subgrade/sdk/jdbc/definition/PhoenixConstants.java
@@ -1,4 +1,4 @@
-package com.dwarfeng.subgrade.sdk.jdbc.database;
+package com.dwarfeng.subgrade.sdk.jdbc.definition;
 
 /**
  * Phoenix 数据库表定义常量。