+++ subgrade-sdk/src/main/java/com/dwarfeng/subgrade/sdk/jdbc/template/GeneralCreateTableTemplate.java
@@ -1,6 +1,6 @@
 package com.dwarfeng.subgrade.sdk.jdbc.template;
 
-import com.dwarfeng.subgrade.sdk.jdbc.database.TableDefinition;
+import com.dwarfeng.subgrade.sdk.jdbc.definition.TableDefinition;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.lang.NonNull;