+++ subgrade-sdk/src/main/java/com/dwarfeng/subgrade/sdk/jdbc/template/PhoenixPresetLookupTemplate.java
@@ -1,11 +1,12 @@
 package com.dwarfeng.subgrade.sdk.jdbc.template;
 
-import com.dwarfeng.subgrade.sdk.jdbc.database.PhoenixUtil;
-import com.dwarfeng.subgrade.sdk.jdbc.database.TableDefinition;
-import com.dwarfeng.subgrade.sdk.jdbc.database.TableDefinitionUtil;
+import com.dwarfeng.subgrade.sdk.jdbc.definition.PhoenixUtil;
+import com.dwarfeng.subgrade.sdk.jdbc.definition.TableDefinition;
+import com.dwarfeng.subgrade.sdk.jdbc.definition.TableDefinitionUtil;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 
 import javax.validation.constraints.NotNull;
+import java.util.Objects;
 
 /**
  * 通用预设查询模板。
@@ -18,9 +19,8 @@ import javax.validation.constraints.NotNull;
  */
 public abstract class PhoenixPresetLookupTemplate extends GeneralPresetLookupTemplate {
 
-    private static final String CACHE_LOOKUP = "lookup";
-    private static final String CACHE_PAGING = "paging";
-    private static final String CACHE_COUNT = "count";
+    private static final String CACHE_SELECT_FROM = "select_from";
+    private static final String CACHE_SELECT_COUNT_FROM = "select_count_from";
 
     public PhoenixPresetLookupTemplate(@NotNull TableDefinition tableDefinition) {
         super(tableDefinition);
@@ -28,45 +28,49 @@ public abstract class PhoenixPresetLookupTemplate extends GeneralPresetLookupTem
 
     @Override
     public String lookupSQL(String preset, Object[] args) {
-        String fragmentSQL;
-        if (containsCachedSQLFragment(CACHE_LOOKUP)) {
-            fragmentSQL = getCachedSQLFragment(CACHE_LOOKUP);
-        } else {
-            fragmentSQL = String.format("SELECT %s FROM %s WHERE %%s",
-                    TableDefinitionUtil.fullColumnSerial(tableDefinition),
-                    TableDefinitionUtil.fullTableName(tableDefinition));
-            putCachedSQLFragment(CACHE_LOOKUP, fragmentSQL);
+        StringBuilder sqlBuilder = new StringBuilder();
+        appendSelectFrom(sqlBuilder);
+        String whereClause = buildWhereClause(preset, args);
+        String orderByClause = buildOrderByClause(preset, args);
+        if (Objects.nonNull(whereClause) && !whereClause.isEmpty()) {
+            sqlBuilder.append(" WHERE ").append(whereClause);
         }
-        return String.format(fragmentSQL, buildWhereClause(preset, args));
+        if (Objects.nonNull(orderByClause) && !orderByClause.isEmpty()) {
+            sqlBuilder.append(" ORDER BY ").append(orderByClause);
+        }
+        return sqlBuilder.toString();
     }
 
     @Override
     public String pagingSQL(String preset, Object[] args, PagingInfo pagingInfo) {
-        String fragmentSQL;
-        if (containsCachedSQLFragment(CACHE_PAGING)) {
-            fragmentSQL = getCachedSQLFragment(CACHE_PAGING);
-        } else {
-            fragmentSQL = String.format("SELECT %s FROM %s WHERE %%s OFFSET %%d LIMIT %%d",
-                    TableDefinitionUtil.fullColumnSerial(tableDefinition),
-                    TableDefinitionUtil.fullTableName(tableDefinition));
-            putCachedSQLFragment(CACHE_PAGING, fragmentSQL);
+        StringBuilder sqlBuilder = new StringBuilder();
+        appendSelectFrom(sqlBuilder);
+        String whereClause = buildWhereClause(preset, args);
+        String orderByClause = buildOrderByClause(preset, args);
+        if (Objects.nonNull(whereClause) && !whereClause.isEmpty()) {
+            sqlBuilder.append(" WHERE ").append(whereClause);
         }
-        return String.format(fragmentSQL, buildWhereClause(preset, args),
-                Math.max(pagingInfo.getPage() * pagingInfo.getRows(), 0), Math.max(pagingInfo.getRows(), 0));
+        if (Objects.nonNull(orderByClause) && !orderByClause.isEmpty()) {
+            sqlBuilder.append(" ORDER BY ").append(orderByClause);
+        }
+        sqlBuilder.append(" OFFSET ").append(Math.max(pagingInfo.getPage() * pagingInfo.getRows(), 0))
+                .append(" LIMIT ").append(Math.max(pagingInfo.getRows(), 0));
+        return sqlBuilder.toString();
     }
 
     @Override
     public String countSQL(String preset, Object[] args) {
-        String fragmentSQL;
-        if (containsCachedSQLFragment(CACHE_COUNT)) {
-            fragmentSQL = getCachedSQLFragment(CACHE_COUNT);
-        } else {
-            fragmentSQL = String.format("SELECT COUNT(%s) FROM %s WHERE %%s",
-                    TableDefinitionUtil.columnSerial(PhoenixUtil.findPrimaryKeyColumns(tableDefinition)),
-                    TableDefinitionUtil.fullTableName(tableDefinition));
-            putCachedSQLFragment(CACHE_COUNT, fragmentSQL);
+        StringBuilder sqlBuilder = new StringBuilder();
+        appendSelectCountFrom(sqlBuilder);
+        String whereClause = buildWhereClause(preset, args);
+        String orderByClause = buildOrderByClause(preset, args);
+        if (Objects.nonNull(whereClause) && !whereClause.isEmpty()) {
+            sqlBuilder.append(" WHERE ").append(whereClause);
         }
-        return String.format(fragmentSQL, buildWhereClause(preset, args));
+        if (Objects.nonNull(orderByClause) && !orderByClause.isEmpty()) {
+            sqlBuilder.append(" ORDER BY ").append(orderByClause);
+        }
+        return sqlBuilder.toString();
     }
 
     /**
@@ -74,11 +78,54 @@ public abstract class PhoenixPresetLookupTemplate extends GeneralPresetLookupTem
      *
      * <p>
      * 例如：
-     * <code>SELECT id, name, age FROM t WHERE <u>name LIKE 'd%' and age > 18<u/></code>
+     * <code>SELECT id, name, age FROM t WHERE <u>name LIKE 'd%' and age > 18</u> ORDER BY age ASC</code>
+     *
+     * <p>
+     * 如果指定的参数没有对应的查询子句，可以直接返回 null。
      *
      * @param preset 查询模板。
      * @param args   查询参数。
      * @return 查询子句。
      */
-    public abstract String buildWhereClause(String preset, Object[] args);
+    protected abstract String buildWhereClause(String preset, Object[] args);
+
+    /**
+     * 构造排序子句。
+     *
+     * <p>
+     * 例如：
+     * <code>SELECT id, name, age FROM t WHERE name LIKE 'd%' and age > 18 ORDER BY <u>age ASC<u/></code>
+     *
+     * <p>
+     * 如果指定的参数没有对应的排序子句，可以直接返回 null。
+     *
+     * @param preset 查询模板。
+     * @param args   查询参数。
+     * @return 排序子句。
+     */
+    protected abstract String buildOrderByClause(String preset, Object[] args);
+
+    private void appendSelectFrom(StringBuilder sqlBuilder) {
+        if (containsCachedSQLFragment(CACHE_SELECT_FROM)) {
+            sqlBuilder.append(getCachedSQLFragment(CACHE_SELECT_FROM));
+        } else {
+            String selectFromFragment = String.format("SELECT %s FROM %s",
+                    TableDefinitionUtil.fullColumnSerial(tableDefinition),
+                    TableDefinitionUtil.fullTableName(tableDefinition));
+            putCachedSQLFragment(CACHE_SELECT_FROM, selectFromFragment);
+            sqlBuilder.append(selectFromFragment);
+        }
+    }
+
+    private void appendSelectCountFrom(StringBuilder sqlBuilder) {
+        if (containsCachedSQLFragment(CACHE_SELECT_COUNT_FROM)) {
+            sqlBuilder.append(getCachedSQLFragment(CACHE_SELECT_COUNT_FROM));
+        } else {
+            String selectFromFragment = String.format("SELECT COUNT(%s) FROM %s",
+                    TableDefinitionUtil.columnSerial(PhoenixUtil.findPrimaryKeyColumns(tableDefinition)),
+                    TableDefinitionUtil.fullTableName(tableDefinition));
+            putCachedSQLFragment(CACHE_SELECT_COUNT_FROM, selectFromFragment);
+            sqlBuilder.append(selectFromFragment);
+        }
+    }
 }